Run genai interface in parallel
