+++ llms/vertexai/internal/common/vertex_client.go
@@ -1,0 +1,131 @@
+package common
+
+import (
+	"context"
+	"fmt"
+	"github.com/tmc/langchaingo/llms/vertexai/internal/aiplatformclient"
+	"github.com/tmc/langchaingo/llms/vertexai/internal/genaiclient"
+	"github.com/tmc/langchaingo/llms/vertexai/internal/schema"
+	"runtime"
+	"strings"
+
+	"google.golang.org/api/option"
+)
+
+const (
+	defaultAPIEndpoint = "us-central1-aiplatform.googleapis.com:443"
+	defaultLocation    = "us-central1"
+	defaultPublisher   = "google"
+)
+
+const (
+	defaultMaxConns = 4
+)
+
+// VertexClient represents a Vertex AI API client.
+type VertexClient struct {
+	genAIClient  genaiclient.GenAIClient
+	legacyClient aiplatformclient.PalmClient
+
+	projectID string
+	location  string
+	publisher string
+}
+
+type googleClient interface {
+	CreateCompletion(context.Context, *schema.CompletionRequest) ([]*schema.Completion, error)
+}
+
+// New returns a new Vertex AI API client.
+func New(ctx context.Context, projectID string, opts ...option.ClientOption) (*VertexClient, error) {
+	numConns := runtime.GOMAXPROCS(0)
+	if numConns > defaultMaxConns {
+		numConns = defaultMaxConns
+	}
+	o := []option.ClientOption{
+		option.WithGRPCConnectionPool(numConns),
+		option.WithEndpoint(defaultAPIEndpoint),
+	}
+	o = append(o, opts...)
+
+	client, err := genaiclient.New(ctx, projectID, defaultLocation, o...)
+	if err != nil {
+		return nil, err
+	}
+
+	legacyClient, err := aiplatformclient.New(ctx, projectID, defaultLocation, o...)
+	if err != nil {
+		return nil, err
+	}
+
+	return &VertexClient{
+		genAIClient:  client,
+		legacyClient: legacyClient,
+		projectID:    projectID,
+		location:     defaultLocation,
+		publisher:    defaultPublisher,
+	}, nil
+}
+
+// CreateCompletion creates a completion.
+func (c *VertexClient) CreateCompletion(ctx context.Context, r *schema.CompletionRequest) ([]*schema.Completion, error) {
+
+	if strings.Contains(r.Model, "bison") || strings.Contains(r.Model, "gecko") {
+		return c.legacyClient.CreateCompletion(ctx, r)
+	}
+
+	return c.genAIClient.CreateCompletion(ctx, r)
+}
+
+// EmbeddingRequest is a request to create an embedding.
+type EmbeddingRequest struct {
+	Input []string `json:"input"`
+	Model string   `json:"model"`
+}
+
+// CreateEmbedding creates embeddings.
+func (c *VertexClient) CreateEmbedding(ctx context.Context, r *EmbeddingRequest) ([][]float32, error) {
+	model := r.Model
+
+	params := map[string]interface{}{}
+	responses, err := c.legacyClient.BatchPredict(ctx, model, r.Input, params)
+	if err != nil {
+		return nil, err
+	}
+
+	embeddings := [][]float32{}
+	for _, res := range responses {
+		value := res.GetStructValue().AsMap()
+		embedding, ok := value["embeddings"].(map[string]interface{})
+		if !ok {
+			return nil, fmt.Errorf("%w: %v", schema.ErrMissingValue, "embeddings")
+		}
+		values, ok := embedding["values"].([]interface{})
+		if !ok {
+			return nil, fmt.Errorf("%w: %v", schema.ErrMissingValue, "values")
+		}
+		floatValues := []float32{}
+		for _, v := range values {
+			val, ok := v.(float32)
+			if !ok {
+				valF64, ok := v.(float64)
+				if !ok {
+					return nil, fmt.Errorf("%w: %v is not a float64 or float32, it is a %T", schema.ErrInvalidValue, "value", v)
+				}
+				val = float32(valF64)
+			}
+			floatValues = append(floatValues, val)
+		}
+		embeddings = append(embeddings, floatValues)
+	}
+	return embeddings, nil
+}
+
+func (c *VertexClient) getModelPath(model string) string {
+	// POST https://{REGION}-aiplatform.googleapis.com/v1/projects/{PROJECT_ID}/locations/{REGION}/publishers/google/models/gemini-pro:streamGenerateContent
+	return fmt.Sprintf("projects/%s/locations/%s/publishers/%s/models/%s", c.projectID, c.location, c.publisher, model)
+}
+
+func (c *VertexClient) CreateChat(ctx context.Context, model string, publisher string, r *schema.ChatRequest) (*schema.ChatResponse, error) {
+	return c.legacyClient.CreateChat(ctx, model, publisher, r)
+}
